fix coin with float number

+++ srcs/coin_order.py
@@ -15,7 +15,7 @@ class coin_order(object):
         self.name = name
         self.symbol = symbol
         self.purchase_price = float(purc_value)
-        self.number = int(number)
+        self.number = float(number)
         self.coin_rank = rank
         self.git_url = git_url
